move route logic into controllers
