+++ platform/controllers/account.js
@@ -1,0 +1,4 @@
+module.exports.accountStatus = (result) => {
+  if(!result) return false
+  return result.beta || false
+}
